requirements.txt
pyyaml
redis
psycopg2-binary
pillow